+++ @xen-orchestra/lite/CHANGELOG.md
@@ -10,6 +10,7 @@
 - [Header] Replace logo with "XO LITE" (PR [#7118](https://github.com/vatesfr/xen-orchestra/pull/7118))
 - New VM console toolbar + Ability to send Ctrl+Alt+Del (PR [#7088](https://github.com/vatesfr/xen-orchestra/pull/7088))
 - Total overhaul of the modal system (PR [#7134](https://github.com/vatesfr/xen-orchestra/pull/7134))
+- More compact UI (PR [#7134](https://github.com/vatesfr/xen-orchestra/pull/7134))
 
 ## **0.1.4** (2023-10-03)